[#737] refactor: 스케줄러 책임 분리 - 스케줄러에 모여있던 부분을 service, eventHandler를 생성하여 책임 분리 - 서비스를 너무 많이 생성하고 service 간에 의존성이 생기긴 함..
